add: CueValidations and CueToJSON
