+++ dag/lib.go
@@ -1,0 +1,5 @@
+package dag
+
+func CreateDag(name string) *Root {
+	return &Root{Name: name}
+}
